Survive stylia deleting matplotlib's font cache
stylia 1.0.1 runs `shutil.rmtree(matplotlib.get_cachedir())` at module scope,
unconditionally. Whichever process imports it first deletes that directory, and
matplotlib does not always recreate it — so the next import raises FileNotFoundError
on '~/.matplotlib', and `olinda validate` or a `learn-soft` figure dies for a font
cache. Hit twice while verifying the report tier, and it would have made the new
report-install CI job intermittently red.

Creating the directory before the import turns the rmtree back into the no-op it
was meant to be. It lives in one function, olinda.style.import_stylia, because the
fix is only worth anything if it is on *every* path that imports stylia; report and
train figures both go through it now, which also folds in the set_style/set_format
call they each made separately.
